requirements.txt
# Dependencias principales del Sistema de Facturación MiChaska
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
reportlab>=4.0.4
Pillow>=10.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0

# Dependencias adicionales para funcionalidades extra
openpyxl>=3.1.0
python-dateutil>=2.8.2
